Move posthog_config context processor to deployment app
Relocate PostHog config resolution out of base and into the
deployment app, which owns the underlying config. Update the
settings registration and move the associated tests.
