+++ config/settings_base.py
@@ -182,7 +182,7 @@ TEMPLATES = [
                 "django.contrib.messages.context_processors.messages",
                 "freedom_ls.site_aware_models.context_processors.site_config",
                 "freedom_ls.accounts.context_processors.signup_policy",
-                "freedom_ls.base.context_processors.posthog_config",
+                "freedom_ls.deployment.context_processors.posthog_config",
                 "django.template.context_processors.csp",
             ],
             "builtins": [
