+++ freedom_ls/base/context_processors.py
@@ -5,7 +5,6 @@ from django.conf import settings
 from django.http import HttpRequest
 
 from freedom_ls.base.git_utils import get_current_branch
-from freedom_ls.deployment.config import config as deployment_config
 
 
 def branch_name_to_color(name: str) -> str:
@@ -41,22 +40,4 @@ def debug_branch_info(_request: HttpRequest) -> dict[str, str]:
         "debug_branch_name": branch,
         "debug_branch_color": bg_color,
         "debug_branch_text_color": get_text_color(bg_color),
-    }
-
-
-def posthog_config(_request: HttpRequest) -> dict[str, str | None]:
-    """
-    Context processor that provides PostHog configuration.
-
-    Args:
-        _request: The current HttpRequest (required by Django context processors)
-
-    Returns:
-        dict: posthog_api_key, posthog_api_host, and posthog_ui_host resolved
-        through freedom_ls.deployment.config.
-    """
-    return {
-        "posthog_api_key": deployment_config.POSTHOG_API_KEY,
-        "posthog_api_host": deployment_config.POSTHOG_API_HOST,
-        "posthog_ui_host": deployment_config.POSTHOG_UI_HOST,
     }
